refactor: move music buttons into subdir
