+++ components/music/buttons/ShareButton/ShareIcon.tsx
@@ -1,6 +1,6 @@
 // Font Awesome Pro 6.3.0 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license (Commercial License) Copyright 2023 Fonticons, Inc.
 
-const Share = () => (
+const ShareIcon = () => (
   <svg
     xmlns="http://www.w3.org/2000/svg"
     viewBox="0 0 512 512"
@@ -12,4 +12,4 @@ const Share = () => (
   </svg>
 );
 
-export default Share;
+export default ShareIcon;
